made fixes based on JKs feedback
