+++ models/claim.js
@@ -1,29 +1,6 @@
 const logger = require('winston');
+const { returnShortId } = require('../helpers/sequelizeHelpers.js');
 const NO_CLAIM = 'NO_CLAIM';
-
-function sortResult (result, longId) {
-  let claimIndex;
-  let shortId = longId.substring(0, 1); // default sort id is the first letter
-  let shortIdLength = 0;
-  // find the index of this certificate
-  claimIndex = result.findIndex(element => {
-    return element.claimId === longId;
-  });
-  if (claimIndex < 0) { throw new Error('claimid not found in possible sorted list') }
-  // get an array of all certificates with lower height
-  let possibleMatches = result.slice(0, claimIndex);
-  // remove certificates with the same prefixes until none are left.
-  while (possibleMatches.length > 0) {
-    shortIdLength += 1;
-    shortId = longId.substring(0, shortIdLength);
-    possibleMatches = possibleMatches.filter(element => {
-      return (element.claimId.substring(0, shortIdLength) === shortId);
-    });
-  }
-  // return the short Id
-  logger.debug('short claim id ===', shortId);
-  return shortId;
-}
 
 module.exports = (sequelize, { STRING, BOOLEAN, INTEGER, TEXT, ARRAY, DECIMAL, DOUBLE, Op }) => {
   const Claim = sequelize.define(
@@ -194,7 +171,7 @@ module.exports = (sequelize, { STRING, BOOLEAN, INTEGER, TEXT, ARRAY, DECIMAL, D
             case 0:
               throw new Error('That is an invalid claim name');
             default:
-              resolve(sortResult(result, claimId));
+              resolve(returnShortId(result, claimId));
           }
         })
         .catch(error => {
